requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.13.4
billiard==4.2.0
blinker==1.8.2
boto3==1.35.1
botocore==1.35.1
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.0
cssselect2==0.7.0
dnspython==2.6.1
docxcompose==1.4.0
docxtpl==0.19.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-async-sqlalchemy==0.6.1
fastapi-cli==0.0.5
fastapi-mail==1.4.1
fastcrud==0.14.0
firebase-admin==6.5.0
flower==2.0.1
fonttools==4.53.1
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.5
grpcio-status==1.65.5
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
humanize==4.10.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.25.1
kaleido==0.2.1
kombu==5.4.0
lxml==5.3.0
Mako==1.3.5
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
numpy==2.1.0
packaging==24.1
pandas==2.2.2
phonenumbers==8.13.43
pillow==10.4.0
plotly==5.23.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pypdf==5.9.0
pyphen==0.16.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-telegram-bot==21.5
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
rich==13.8.0
rsa==4.9
s3transfer==0.10.2
sentry-sdk==2.13.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
strapi-client==2.9.1
tenacity==9.0.0
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.23.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
websockets==12.0
yarl==1.9.4
zopfli==0.2.3
